registration/login(the correct one with Long type for id)

+++ src/main/java/com/bootcamp/interviewflow/dto/UserResponseDto.java
@@ -3,25 +3,25 @@ package com.bootcamp.interviewflow.dto;
 import java.time.LocalDateTime;
 
 public class UserResponseDto {
-    private Integer id;
+    private Long id;
     private String name;
     private String email;
     private LocalDateTime createdAt;
 
     public UserResponseDto() {}
 
-    public UserResponseDto(Integer id, String name, String email, LocalDateTime createdAt) {
+    public UserResponseDto(Long id, String name, String email, LocalDateTime createdAt) {
         this.id = id;
         this.name = name;
         this.email = email;
         this.createdAt = createdAt;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
